Warnings building with /W4 on Visual Studio 2013; Project files changed to Visual Studio 2013, but should load normally on 2012, so I didn't bothered to fix/duplicate; Fix workdir for VS projects; Fix lib dir for VS projects; Fix prefab problem when no prefab was used; Some other random fixes;

+++ samples/06_sprite/sprite_sample.cpp
@@ -91,8 +91,8 @@ void SpriteSample::OnInputPointerRelease(const EventInputPointer *ev)
 		if (pSprite)
 			vFrom = pSprite->GetPosition();
 
-		vTo.setX(ev->GetX());
-		vTo.setY(ev->GetY());
+		vTo.setX(f32(ev->GetX()));
+		vTo.setY(f32(ev->GetY()));
 		vTo += pCamera->GetPosition();
 		fElapsed = 0.0f;
 	}
